Fixed some conversion to KB errors and developed a fix for reclaim
The fix for reclaim hinges on using a different type of message for when a putchunk is sent in response of a "REMOVED" message being received. It is a new case of message "PUTCHUNKREMOVED", in this case, the peer will verify if it has that chunk in it's blacklist. The blacklist is made as the peer deletes chunks to open up space for space reclaim. However, this means that it will never be able to backup those chunks if it receives a PUTCHUNKREMOVED. This is problematic because if the same chunks are reclaimed in another peer than it won't be able to back them up if it receives a PUTCHUNKREMOVED from even another peer that intercepted the first one's REMOVED message. We need to find a way to make it so it "knows" that it is not supposed  to accept back it's own chunks in that way.